==> 01-setup.Rmd <==

# Setup {#Setup}

```{r include=FALSE}
# automatically create a bib database for R packages
knitr::write_bib(c(
  .packages(), 'bookdown', 'knitr', 'rmarkdown'
), 'packages.bib')
```

## Workspace setup {#Workspace_setup}

We specify the folder where the analysis is located. 

```{r}
setwd("~/projects/oligomm-claudia/")
```

## Dependencies

Functions that are used frequent are defined in `utils.R` and are mostly helper functions for data manipulation (e.g. translation of sample ids to groups). Other packages can be load via cran using the `install.packages()` function.

```{r, results="hide"}
rm(list = ls())
source("utils.R")
library("vcfR")
library("rmdformats")
library("ggplot2")
library("reshape2")
library("ggridges")
library("knitr")
library("kableExtra")
library("plotly")
library("DT")
library("rtracklayer")
```

 for `rtracklayer` you might need curl
```{bash, eval = F}
sudo apt-get install libcurl4-openssl-dev
```

## Data structure {#Data_structures}

`data/raw`
: files from host machine, contains `coverage`, _lofreq_ and _varscan_ output

`data/annotation`
: annotation files for OMM12 genomes e.g. phage location (*pro-hunter*, *fasta* files and *Prokka* annotation)

`data/tables`
: tables in tsv format for e.g. import with Excel

`data/annotation/gff`
: *Prokka* annotation

`data/rda`
: processed raw data in *rds* format generated by `01-load-data.rmd`

`docs`
:  `html` files generated via `bookdown::serve_book(getwd())`
