==> 03-variant-statistics.Rmd <==
# Variant statistics {#Variant_statistics}

## Process of _raw_ data

Since the raw output of the **Snakemake** pipeline produces one `vcf` file per sample (including all contigs of the OligoMM reference genomes) we now produce a single `rds` file that contains the merged variant information of the whole study.

We define a function that iterates over a set of `vcf` files (`vcf.files`)

```{r}
#' Processes a list of vcf files
#' 
#' @param vcf.files a list of file path of files in vcf format
#' @param contig_mapping mapping file of vcf CHR to OMM reference genomes
#' @param gff.dff merged gff data from reference genomes
#' @return a dataframe holding the merged annotated variant information
vcfToDataframe <- function(vcf.files, contig_mapping = read.csv2("data/contig_genome_mapping.csv"), gff.df = gff.df, varscan = F) {
	require(vcfR)
	res <- list()
	for (file in vcf.files) {
		message(file)
		vcf.content <- vcfR::read.vcfR(file, verbose = FALSE )
		vcf.fix <- as.data.frame(vcf.content@fix) # contains chr, position and substitution informations
		vcf.info <- vcfR::INFO2df(vcf.content) # get INFO field, contains DP, AF informations
	  vcf.ann <- data.frame(do.call('rbind', strsplit(as.character(vcf.info$ANN),'|',fixed = TRUE))) # split ANN field, contains information if files are processed by snpEff
		if(nrow(vcf.fix) > 0) {
			if (varscan){
				dat <-  as.data.frame(cbind(vcf.fix[,c(1, 2, 4, 5, 6)], vcf.info[,c(1, 2, 3)])) # ther is a PV column only present in varscan files
				dat$pvalue <- as.numeric(as.matrix(vcf.info$PV))

			} else {
				dat <-  as.data.frame(cbind(vcf.fix[,c(1, 2, 4, 5, 6)], vcf.info[,c(1, 2)]))
			}
	
			dat$majorAF <- sapply(dat$AF, minorAfToMajorAf) # transforms e.g. AF of 0.1 to 0.9, 0.9 stays 0.9 and 0.5 stays 0.5
			dat$genome <- contig_mapping[match(dat$CHROM, contig_mapping$contig),]$genome # map chr information to genome name e.g. NHMU01000001.1 -> i48
			dat$genome_hr <- translateGenomeIdToFullName(dat$genome)
			dat$mouse.id <- substr(basename(file), 1, 4) # get mouse ID from file name, e.g. "1681"
			dat$dp <- as.numeric(as.matrix(vcf.info$DP))
		#	print(summary(dat$AF))
		#	print(summary(dat$DP))
			dat$vartype <- as.character(as.matrix(vcf.ann$X2))
			dat$effect <- as.character(as.matrix(vcf.ann$X3))
			dat$substitution <- as.character(as.matrix(vcf.ann$X11))
			# annotate overlay of gene
			dat$orf_desc <- as.character(as.matrix(apply(dat, 1, getGeneByPosition, gff = gff, pos.column = 2, chr.column = 1)))
			res[[basename(file)]] <- dat # add vcf df to list
		} else{
			message("Skipping")
		}
	}
	df <- as.data.frame(do.call(rbind, res)) # merge list to df
	return(df)
}
```

We apply this function on all  `r length(Sys.glob("data/raw/lofreq/*.vcf"))` files that live in `data/raw/lofreq/*.vcf`

```{r, results="hide", message=FALSE, eval = F}
# merge gff annotaions
gff.files <- Sys.glob("data/annotation/gff/*.gff")
gff.df <- NULL
for (gff.file in gff.files){
	message(gff.file)
	gff <- rtracklayer::readGFF(gff.file)
	# subset since different columns are present on gff files
	relevant <- data.frame(start = gff$start, end = gff$end, type = as.character(as.matrix(gff$type)), gene = as.character(as.matrix(gff$gene)), product = as.character(as.matrix(gff$product)), chr = as.character(as.matrix(gff$seqid)))
	relevant$genome <-  substr(basename(gff.file), 1, nchar(basename(gff.file))-4)
	gff.df <- rbind(gff.df, relevant)
}

contig_mapping <- read.csv2("data/contig_genome_mapping.csv", sep =";", header=T, stringsAsFactors = F) # this file contains contig names of the 12 OligoMM genomes

# for lofreq results
vcf.files <- Sys.glob("data/raw/lofreq_eff/*.vcf")
vcf.samples <- suppressWarnings(vcfToDataframe(vcf.files, contig_mapping, gff.df = gff.df))
# annotate study metadata
vcf.samples <- merge(vcf.samples, design.df, by = "mouse.id")
vcf.samples$mouse.id <- as.character(as.matrix(vcf.samples$mouse.id))
vcf.samples[which(vcf.samples$orf_desc == "character(0)"),]$orf_desc <- "outside ORFs"
vcf.samples[which(vcf.samples$orf_desc == "hypothetical protein"),]$orf_desc <- "hypothetical proteins"
vcf.samples[which(is.na(vcf.samples$vartype)),]$vartype <- "unknown"
saveRDS(vcf.samples, file = "data/rds/variants_lofreq.rds")

# for varscan results
vcf.files <- Sys.glob("data/raw/varscan_eff/*.vcf")
vcf.samples <- suppressWarnings(vcfToDataframe(vcf.files, contig_mapping, gff.df = gff.df, varscan = T))
# annotate study metadata
vcf.samples <- merge(vcf.samples, design.df, by = "mouse.id")
vcf.samples$mouse.id <- as.character(as.matrix(vcf.samples$mouse.id))
vcf.samples[which(vcf.samples$orf_desc == "character(0)"),]$orf_desc <- "outside ORFs"
vcf.samples[which(vcf.samples$orf_desc == "hypothetical protein"),]$orf_desc <- "hypothetical proteins"
vcf.samples[which(is.na(vcf.samples$vartype)),]$vartype <- "unknown"
saveRDS(vcf.samples, file = "data/rds/variants_varscan.rds")
```

The dataset is written to `data/rds/variants_lofreq.rds` and `data/rds/variants_varscan.rds`

## Quality of variants

DP is the filtered depth, at the sample level. This gives you the number of filtered reads that support each of the reported alleles. You can check the variant caller’s documentation to see which filters are applied by default. Only reads that passed the variant caller’s filters are included in this number. However, unlike the AD calculation, uninformative reads are included in DP.

### lofreq

```{r load-vcf-qual-lo, fig.cap="quality of variants (all samples)", fig.align="center", fig.width=15}
dat <- readRDS("data/rds/variants_lofreq.rds")
dat <- dat[which(dat$ecoli== FALSE),]
print(summary(dat$dp))
```

### all samples

```{r plot-dp-all, fig.cap="filtred depth of varaints (all samples)", fig.align="center", fig.width=10}
p <- ggplot(dat, aes(dp, fill = genome)) + geom_histogram()
p <- p + theme_classic() + xlab("filtered depth") + ylab("occurence") 
p <- p + scale_fill_manual(values = omm_colors)
plotly::ggplotly(p)
```
### by mouse

```{r plot-dp-by-mouse, fig.cap="filtred depth of varaints by mouse", fig.height = 20, fig.width = 10}
p <- ggplot(dat, aes(dp, fill = genome)) + geom_histogram()
p <- p + theme_classic() + xlab("filtered depth") + ylab("occurence") 
p <- p + scale_fill_manual(values = omm_colors) + facet_wrap(~ mouse.id + desc, ncol = 2)
p <- p + theme(panel.border=element_blank(),
                 strip.text=element_text(size=12, colour="black"),
                 strip.background=element_rect(colour="white", 
                                               fill="white"))
plotly::ggplotly(p)
```

## Number of variants 

### by sample

only non-ecoli samples

```{r load-vcf}
dat <- readRDS("data/rds/variants_lofreq.rds")
dat <- dat[which(dat$ecoli== FALSE),]
```

This set contains `r nrow(dat)` variants found in `r length(unique(dat$genome))` genomes. The detailed breakdow of variants per genome is as follows.

```{r}
dat$dummy <- 1
dat.by.sample.genome <- aggregate(dummy ~ mouse.id + genome, dat, sum)
colnames(dat.by.sample.genome) <- c("sample", "genome", "number_variants")
DT::datatable(dat.by.sample.genome)
```

```{r, fig.cap="Number of variants by genome (color) found in samples"}
p <- ggplot(dat.by.sample.genome, aes(x = reorder(sample, number_variants), y = number_variants, fill = genome))
p <- p + geom_bar(stat="identity") + coord_flip() + theme_classic()
p <- p + ylab("number of variants") + xlab("sample") + scale_fill_manual(values = omm_colors)
plotly::ggplotly(p)
```

### by mouse and genome

only non-ecoli samples

```{r}
dat$dummy <- 1
dat.by.sample <- aggregate(dummy ~ mouse.id, dat, sum)
colnames(dat.by.sample) <- c("sample", "number_variants")
DT::datatable(dat.by.sample)
```

```{r, fig.align="center", fig.width=10, fig.cap="number of variants of all 12 OMM genomes by mouse"}
p <- ggplot(dat.by.sample, aes(x = reorder(sample, number_variants), y = number_variants))
p <- p + geom_bar(stat="identity") + coord_flip() + theme_classic()
p <- p + xlab("number of variants") + ylab("sample")
plotly::ggplotly(p)
```

### by extendet metadata

only non-ecoli samples

```{r}
dat$dummy <- 1
dat.by.sample.genome2 <- aggregate(dummy ~ mouse.id + genome + generation + ecoli + desc + day, dat, sum)
colnames(dat.by.sample.genome2) <- c("sample", "genome", "generation", "ecoli", "desc","day","number_variants")
```

```{r}
DT::datatable(dat.by.sample.genome2)
dat$DP <- NULL
```

```{r, fig.align="center", fig.width=10, fig.cap="Number of variants by genome (color) found in samples"}
#dat.by.sample.genome2 <- dat.by.sample.genome2[which(dat.by.sample.genome2$sample != "I_cc" & dat.by.sample.genome2$sample != "I_mi"),]
p <- ggplot(dat.by.sample.genome2, aes(x = sample, y = number_variants, fill = genome))
p <- p + geom_bar(stat="identity") + theme_classic()
p <- p + facet_grid(.~ generation + day, scales = "free_x")
p <- p + ylab("number of variants") + xlab("sample") + scale_fill_manual(values = omm_colors)
p <- p + theme(panel.border=element_blank(),
                 strip.text=element_text(size=12, colour="black"),
                 strip.background=element_rect(colour="white", 
                                               fill="white"))
print(p) # plotly is not working on this type of plot
```

## ORFs with variants

In this section we visualze the occurence of ORFs and their annotations of overlaying variants. The full table is available in this github repository under `data/tables/table1.tsv`.

### Data

#### Write to table

```{r}
write.table(dat, file = "data/tables/table1.tsv", sep = "\t", row.names = F, quote = F)
```

#### Aggregate by mouse

The total number of variant overlapts to ORF per mouse, ordered by their occurence is below and can in  `data/tables/table2.tsv`

```{r}
dat.all <- aggregate(dummy ~ genome + orf_desc , dat, sum)
colnames(dat.all) <- c("genome", "orf_annotation", "occurence")
write.table(dat.all, file = "data/tables/table2.tsv", sep = "\t", row.names = F, quote = F)
dat.all  <- dat.all[order(dat.all$occurence, decreasing = TRUE),]
DT::datatable(dat.all)
```

### Statistics

In total, we identified `r nrow(dat)` variants in all `r length(unique(dat$mouse.id))` samples. Of them, `r sum(dat.all$occurence)` variants map to `r length(unique(dat.all$orf_annotation))` ORFs with an annotated function

### Minor to major transistion

#### Plotting function

```{r}
pairedAFplot <- function(dat = dat, case = "1607", control = "1885"){
	require(reshape2) 
	require(ggplot2)
	dat.pair <- dat[which(dat$mouse.id == case | dat$mouse.id == control),]
	#dat.pair <- dat.pair[which(!is.na(dat.pair$orf_desc)),]
	data.wide <- dcast(dat.pair, genome + POS + REF + ALT + orf_desc + vartype ~ mouse.id, value.var = "AF")
	colnames(data.wide)[7] <- "sample_1"
	colnames(data.wide)[8] <- "sample_2"
	data.wide[is.na(data.wide)] <- 0

	p <- ggplot(data.wide, aes(x = sample_1, y =sample_2, color = genome, label = orf_desc))#, shape = vartype))
	p <- p + geom_point(fill = NA) + theme_classic() + scale_color_manual(values = omm_colors)
	p <- p + geom_hline(yintercept = .5, linetype = 2, color = "grey50")
	p <- p + geom_vline(xintercept = .5, linetype = 2, color = "grey50")
	p <- p + theme(aspect.ratio = 1) 
	#p <- p + scale_shape_manual(values = variant_shapes)
	p <- p + theme(strip.background = element_blank())
	p <- p + geom_abline(intercept = 0, slope = 1)
	p <- p + xlab(paste0("sample ", case)) + ylab(paste0("sample ", control))
	p <- p + theme(panel.border = element_rect(colour = "black", fill=NA, size=1))
	return(p)
}
```

#### between generation comparison

##### comparison of t=20 of first/second generation

```{r a1, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1607", control = "1885")))
```

##### comparison of t=40 of first/second generation (sample 1660)

```{r a2, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1660", control = "1807")))
```

##### comparison of t=40 of first/second generation (sample 1750)

```{r a3, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1750", control = "1807")))
```

##### comparison of t=80 of first/second generation

```{r a4, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1779", control = "1789")))
```

##### comparison of t=80 of first/second generation

```{r a5, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1779", control = "1789")))
```

#### within generation comparison of gen=1

##### t=20 vs t=40

```{r a6, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1607", control = "1660")))
```
##### t=40 vs t=80

```{r a7, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1660", control = "1779")))
```

#### within generation comparison of gen=2

##### t=20 vs t=40

```{r a8, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1885", control = "1807")))
```
##### t=40 vs t=80

```{r a9, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "1807", control = "1789")))
```

#### I_mi comparison

##### t=20 gen=2

 ```{r a10, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "I_mi", control = "1885")))
```

##### t=40 gen=2

 ```{r a, fig.align="center", fig.width=10, fig.height=10, fig.cap="Scatter plot of AF of two samples. Only variants with functional annotation are shown"}
suppressWarnings(ggplotly(pairedAFplot(dat, case = "I_mi", control = "1807")))
```

### Gene-centric analysis

#### function

```{r}
geneCentricPlot <- function(dat = dat, genome = "yl44"){
	require(reshape2) 
	require(ggplot2)
	dat.subset <- dat[which(dat$genome == genome),]
	dat.subset$ID <- paste0(dat.subset$POS, "|", dat.subset$REF ,"->", dat.subset$ALT)
	p <- ggplot(dat.subset, aes(x = ID, y = AF, color = desc, shape = vartype))
  p <- p + geom_point(size = 4, fill = NA)
	p <- p + coord_flip() + theme_classic()
	p <- p + facet_grid(orf_desc ~., scales = "free", space = "free")
	p <- p + ylim(c(0,1))
#	p <- p + scale_shape_manual(values = variant_shapes)
	p <- p + theme(strip.text.y = element_text(angle = 0), strip.placement = "outside")
	p <- p + xlab("") + ylab("allele frequency")  + scale_color_manual(values = sample_colors)
	return(p)
}

# no grid version for html
geneCentricPlot2 <- function(dat = dat, genome = "yl44"){
	require(reshape2) 
	require(ggplot2)
	dat.subset <- dat[which(dat$genome == genome),]
	dat.subset$ID <- paste0(dat.subset$POS, "|", dat.subset$REF ,"->", dat.subset$ALT)
	p <- ggplot(dat.subset, aes(x = orf_desc, y = AF, color = desc, shape = vartype))
  p <- p + geom_point(size = 4, fill = NA)
	p <- p + coord_flip() + theme_classic()
	p <- p + ylim(c(0,1))
	#p <- p + scale_shape_manual(values = variant_shapes)
	p <- p + theme(strip.background = element_blank()) + theme(strip.text.y = element_text(angle = 0), strip.placement = "outside")
	p <- p + xlab("") + ylab("allele frequency") + scale_color_manual(values = sample_colors)
	return(p)
}
```

#### by ORF annotation (html)

##### yl44

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "yl44")))
```

##### yl27

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "yl27")))
```

##### yl58

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "yl58")))
```

##### i48

```{r, fig.align="center", fig.width=15, fig.height=15, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "i48")))
```

##### yl32

```{r, fig.align="center", fig.width=15, fig.height=15, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "yl32")))
```

##### i46

```{r, fig.align="center", fig.width=15, fig.height=15, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "i46")))
```

##### yl45

```{r, fig.align="center", fig.width=15, fig.height=15, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "yl45")))
```

##### yl31

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "yl31")))
```

##### yl2

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "yl2")))
```

##### i49

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "i49")))
```

##### kb18

no variant

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
#suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "kb18"))
```

##### kb1

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(ggplotly(geneCentricPlot2(dat, genome = "kb1")))
```

### by variant (static)

similar to the plots above, but now the y axis is not the annotation but the variant ID (e.g. `position|reference->alternative`)

##### yl44

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "yl44"))
```

##### yl27

```{r, fig.align="center", fig.width=15, fig.height=8, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "yl27"))
```

##### yl58

```{r, fig.align="center", fig.width=15, fig.height=18, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "yl58"))
```

##### i48

```{r, fig.align="center", fig.width=15, fig.height=18, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "i48"))
```

##### yl32

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "yl32"))
```

##### i46

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "i46"))
```

##### yl45

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "yl45"))
```

##### yl31

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "yl31"))
```

##### yl2

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "yl2"))
```

##### i49

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "i49"))
```

##### kb18

no variant

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
#suppressWarnings(ggplotly(geneCentricPlot(dat, genome = "kb18"))
```

##### kb1

```{r, fig.align="center", fig.width=15, fig.height=10, fig.cap="AF by function of one genome"}
suppressWarnings(geneCentricPlot(dat, genome = "kb1"))
```
